+++ packaging/bm-eas/redhat/bm-eas.spec
@@ -32,26 +32,28 @@ install -m 644 /sources/stretch/bm-eas.service %{buildroot}%{_unitdir}
 %pre
 if [ $1 -gt 1 ]; then
     # Upgrade
-    systemctl stop bm-eas
+    [ -d /run/systemd/system ] && systemctl stop bm-eas
 fi
 
 %post -p /bin/bash
-systemctl daemon-reload
 systemctl enable bm-eas
+if [ -d /run/systemd/system ]; then
+    systemctl daemon-reload
 
-if [ $1 -eq 1 ]; then
-    # Installation
-    systemctl start bm-eas
+    if [ $1 -eq 1 ]; then
+        # Installation
+        systemctl start bm-eas
+    fi
 fi
 
 %preun
 if [ $1 -eq 0 ]; then
     # Uninstall
-    systemctl stop bm-eas
+    [ -d /run/systemd/system ] && systemctl stop bm-eas
 fi
 
 %postun
 if [ $1 -eq 1 ]; then
-    # Upgrade
-    systemctl start bm-eas
+    # Upgrade
+    [ -d /run/systemd/system ] && systemctl start bm-eas
 fi
