+++ packaging/bm-node/redhat/bm-node.spec
@@ -34,26 +34,28 @@ mkdir -p %{buildroot}/var/lib/bm-ca
 %pre
 if [ $1 -gt 1 ]; then
     # Upgrade
-    systemctl stop bm-node
+    [ -d /run/systemd/system ] && systemctl stop bm-node
 fi
 
 %post -p /bin/bash
-systemctl daemon-reload
 systemctl enable bm-node
+if [ -d /run/systemd/system ]; then
+    systemctl daemon-reload
 
-if [ $1 -eq 1 ]; then
-    # Installation
-    systemctl start bm-node
+    if [ $1 -eq 1 ]; then
+        # Installation
+        systemctl start bm-node
+    fi
 fi
 
 %preun
 if [ $1 -eq 0 ]; then
     # Uninstall
-    systemctl stop bm-node
+    [ -d /run/systemd/system ] && systemctl stop bm-node
 fi
 
 %postun
 if [ $1 -eq 1 ]; then
-    # Upgrade
-    systemctl start bm-node
+    # Upgrade
+    [ -d /run/systemd/system ] && systemctl start bm-node
 fi
