+++ packaging/bm-tika/redhat/bm-tika.spec
@@ -33,26 +33,27 @@ install -m 644 /sources/stretch/bm-tika.service %{buildroot}%{_unitdir}
 %pre
 if [ $1 -gt 1 ]; then
     # Upgrade
-    systemctl stop bm-tika
+    [ -d /run/systemd/system ] && systemctl stop bm-tika
 fi
 
 %post -p /bin/bash
-systemctl daemon-reload
 systemctl enable bm-tika
-
-if [ $1 -eq 1 ]; then
-    # Installation
-    systemctl start bm-tika
+if [ -d /run/systemd/system ]; then
+    systemctl daemon-reload
+    if [ $1 -eq 1 ]; then
+        # Installation
+        systemctl start bm-tika
+    fi
 fi
 
 %preun
 if [ $1 -eq 0 ]; then
     # Uninstall
-    systemctl stop bm-tika
+    [ -d /run/systemd/system ] && systemctl stop bm-tika
 fi
 
 %postun
 if [ $1 -eq 1 ]; then
-    # Upgrade
-    systemctl start bm-tika
+    # Upgrade
+    [ -d /run/systemd/system ] && systemctl start bm-tika
 fi
