+++ packaging/bm-xivobridge/redhat/bm-xivobridge.spec
@@ -35,26 +35,29 @@ install -m 644 /sources/stretch/bm-xivobridge.service %{buildroot}%{_unitdir}
 %pre
 if [ $1 -gt 1 ]; then
     # Upgrade
-    systemctl stop bm-xivobridge
+    [ -d /run/systemd/system ] && systemctl stop bm-xivobridge
 fi
 
 %post -p /bin/bash
-systemctl daemon-reload
 systemctl enable bm-xivobridge
 
-if [ $1 -eq 1 ]; then
-    # Installation
-    systemctl start bm-xivobridge
+if [ -d /run/systemd/system ]; then
+    systemctl daemon-reload
+
+    if [ $1 -eq 1 ]; then
+        # Installation
+        systemctl start bm-xivobridge
+    fi
 fi
 
 %preun
 if [ $1 -eq 0 ]; then
     # Uninstall
-    systemctl stop bm-xivobridge
+    [ -d /run/systemd/system ] && systemctl stop bm-xivobridge
 fi
 
 %postun
 if [ $1 -eq 1 ]; then
-    # Upgrade
-    systemctl start bm-xivobridge
+    # Upgrade
+    [ -d /run/systemd/system ] && systemctl start bm-xivobridge
 fi
