+++ packaging/bm-xmpp/redhat/bm-xmpp.spec
@@ -31,26 +31,28 @@ install -m 644 /sources/stretch/bm-xmpp.service %{buildroot}%{_unitdir}
 %pre
 if [ $1 -gt 1 ]; then
     # Upgrade
-    systemctl stop bm-xmpp
+    [ -d /run/systemd/system ] && systemctl stop bm-xmpp
 fi
 
 %post -p /bin/bash
-systemctl daemon-reload
 systemctl enable bm-xmpp
+if [ -d /run/systemd/system ]; then
+    systemctl daemon-reload
 
-if [ $1 -eq 1 ]; then
-    # Installation
-    systemctl start bm-xmpp
+    if [ $1 -eq 1 ]; then
+        # Installation
+        systemctl start bm-xmpp
+    fi
 fi
 
 %preun
 if [ $1 -eq 0 ]; then
     # Uninstall
-    systemctl stop bm-xmpp
+    [ -d /run/systemd/system ] && systemctl stop bm-xmpp
 fi
 
 %postun
 if [ $1 -eq 1 ]; then
-    # Upgrade
-    systemctl start bm-xmpp
+    # Upgrade
+    [ -d /run/systemd/system ] && systemctl start bm-xmpp
 fi
